iter2: Move user API handlers
