+++ internal/api/handler/user/register/handler.go
@@ -1,51 +1,51 @@
-package login
+package register
 
 import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/bjlag/go-loyalty/internal/usecase/user/register"
 	"net/http"
 
 	"github.com/bjlag/go-loyalty/internal/infrastructure/logger"
-	"github.com/bjlag/go-loyalty/internal/usecase/login"
 )
 
 type Handler struct {
-	usecase *login.Usecase
+	usecase *register.Usecase
 	log     logger.Logger
 }
 
-func NewHandler(usecase *login.Usecase, log logger.Logger) *Handler {
+func NewHandler(usecase *register.Usecase, log logger.Logger) *Handler {
 	return &Handler{
 		usecase: usecase,
 		log:     log,
 	}
 }
 
-func (h *Handler) Handle(w http.ResponseWriter, r *http.Request) {
+func (h Handler) Handle(w http.ResponseWriter, r *http.Request) {
 	var req Request
 
 	err := json.NewDecoder(r.Body).Decode(&req)
 	if err != nil {
 		if errors.Is(err, errInvalidLogin) || errors.Is(err, errInvalidPassword) {
-			h.log.WithError(err).Warning("Invalid request")
+			h.log.WithError(err).Warning("invalid request")
 			http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
 			return
 		}
 
-		h.log.WithError(err).Error("Error decoding request")
+		h.log.WithError(err).Error("error decoding request")
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 		return
 	}
 
-	token, err := h.usecase.LoginUser(r.Context(), req.Login, req.Password)
+	token, err := h.usecase.RegisterUser(r.Context(), req.Login, req.Password)
 	if err != nil {
-		if errors.Is(err, login.ErrUserNotFound) || errors.Is(err, login.ErrWrongPassword) {
-			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
+		if errors.Is(err, register.ErrUserAlreadyExists) {
+			http.Error(w, http.StatusText(http.StatusConflict), http.StatusConflict)
 			return
 		}
 
-		h.log.WithError(err).Error("Error login user")
+		h.log.WithError(err).Error("error registering user")
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 		return
 	}
@@ -56,7 +56,7 @@ func (h *Handler) Handle(w http.ResponseWriter, r *http.Request) {
 
 	data, err := json.Marshal(resp)
 	if err != nil {
-		h.log.WithError(err).Error("Could not marshal response")
+		h.log.WithError(err).Error("could not marshal response")
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 		return
 	}
@@ -66,7 +66,7 @@ func (h *Handler) Handle(w http.ResponseWriter, r *http.Request) {
 
 	_, err = w.Write(data)
 	if err != nil {
-		h.log.WithError(err).Error("Could not write response")
+		h.log.WithError(err).Error("could not write response")
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
 	}
 }
